разбивает js на модули
